+++ docs/source/changelog.rst
@@ -56,7 +56,8 @@ v26.1.9-dev
 - Added :class:`~discord.events.GuildRoleUpdateEvent` event.
 - Added :class:`~discord.events.GuildRoleDeleteEvent` event.
 - Added :class:`~discord.events.GuildScheduledEventUserAddEvent` event.
-- Added :class:`~dsicord.events.GuildScheduledEVentUserRemoveEvent` event.
+- Added :class:`~discord.events.GuildScheduledEVentUserRemoveEvent` event.
+- Added :class:`~discord.events.GuildSoundboardSoundDeleteEvent` event.
 
 
 **Documentation Changes**